Support setting filter values in URL params on the individual query page
This is useful for linking someone directly to a specific query with
preset filters.  It's only done on first load so as not to interfere
with interactive use such as switching filter values or updating the
dataset.

It would also be nice if current filter values were preserved across
dataset updates, but I think that's best done as a separate change to a
different part of the code.  (Perhaps QueryResult.getQueryResult()?  I'm
not familiar enough with the code yet to know.)