Remove webpack from Docker Compose:
I realized that it makes more sense to run Webpack on the host machine
as usually you will need to install Node's tools on the host to enjoy
better intelisense and eslint support in the editor.

On top of that, installing Node.js on most developer machines should be
relatively easy.