Disable version counter for queries:
Since the move to SQLAlchemy, version is managed automatically.
When running a query, the worker updates the query row and increments
the version. This results in the user's inability to save the query
after that.

From reviewing SQLA's documentation, it looks like we will have to
manage the version value ourselves. For now, I'm just disabling the
version auto increment. In a follow up update we need to actually
implement the manual increment where we think it's needed.