Enforce only LDAP Login
When enabling LDAP login, users are still shown with the initial login
screen and with a link they have to click to use LDAP Auth.

We want to completely switch to LDAP authentication and in order to do
this, we need a new env variable: LDAP_ONLY_LOGIN and a minor change in
ldap/authentication/__init__.py to redirect the user directly to the
ldap login page.