feat: support hiding Redshift tables from schema list
These changes make it possible to configure a Redshift data source so
that certain tables are hidden from users in the schema browser.

This feature is useful because the Redshift SVV_COLUMNS table used to
retrieve the schema returns schema for every single table in the
database. This means that the schema for a Redshift data source also
includes tables that cannot be accessed from Redash. This clutters the
schema browser and confuses users of Redshift data sources.

The exclude filter can be used to manually hide the tables that are not
accessible from Redash. The filter is evaluated as a regular expression
which makes it rather flexible for various cases.