when a normal BigQuery query fails, it returns a 4xx code and an explicit error message.
BigQuery recently added scripting capabilities to their queries. These queries behave a bit differently, as they spawn child jobs.

Due to this change, when a scripting query fails, it queryResults no longer returns a 4xx code, but instead returns a 200 with an incomplete JSON. This causes our code normal to  when trying to  get the 'schema' node and result rows.

This commit fixes that, so that if there is no 'schema' row, it will get more details about the failed job, details that will explicitly contain the error, if any.