Sort numeric values as numbers rather than strings
The way things currently work, if a Category-type axis has numeric values with differing numbers of digits (e.g. [1 2 10]) and you sort it, you'll end up with values that don't seem properly (numerically-speaking) sorted (e.g. [1 10 2]). This PR changes that so that numeric strings are sorted as numbers.

If you were to have a mix of numeric and non-numeric strings, we currently return the numbers first. This behavior is preserved by using Infinity as the "parsed" value for non-numeric strings. Non-numeric strings are then properly sorted via the same _.identity method as currently.