Change new data source group attachment
As both data sources and  new users are automatically added to the default group (at least when using SAML authentication), the creation of more groups would be silly as everyone would have access to all data sources anyway. So an idea would be to attach new data sources to the admin group instead.