Update Cypress to work without Hipchat
This commit fixes a weird looking Cypress error:

    expected [ Array(8) ] to not be a superset of []

After investigation, that error was because Hipchat was the only
"deprecated" destination defined in Redash.  So removing it made
for an empty array, which the old "not.to.contain.members" test
can't handle.  Thus we use "not.to.contain.oneOf", which can.