Support a python-table-object string in query_Results query-runner
Allow to add a table in python-string format in the sqlite command to extend the possibilities of the query_results, especially in combination with the python query runner.

the table-string has to be preceded by "query_table_start_" and followed by "_query_table_end"

This is most likely not a safe feature. To enable it, the requirement variable REDASH_QUERY_RESULTS_ALLOW_PYTHON_TABLE_STRING has to be set to "true".