fix: Really fix alembic multiple heads issue
The previous commit was the solution used in the last release, and
apparently worked once as it was able to find one upgrade, but fails a
second time when it can't find any upgrades despite there being multiple
heads in alembic_version.

Retrofitting Alembic branches into a project that doesn't already have
them is annoying, so we should in future upgrades just manually turn
merge any stacklet versions in a new upstream version. This will keep
alembic_version down to a single row and de-confuse upgrades.